tests/wpt/tests/shared-storage/run-url-selection-operation.tentative.https.sub.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/shared-storage/resources/util.js"></script>
<script src="/fenced-frame/resources/utils.js"></script>

<body>
<script>
'use strict';

for (const resolve_to_config of [true, false]) {
  promise_test(async () => {
    const ancestor_key = token();
    let url0 = generateURL("/shared-storage/resources/frame0.html",
                           [ancestor_key]);
    let url1 = generateURL("/shared-storage/resources/frame1.html",
                           [ancestor_key]);

    await addModuleOnce("/shared-storage/resources/simple-module.js");

    let select_url_result0 = await sharedStorage.selectURL(
        "test-url-selection-operation", [{url: url0}, {url: url1}],
        {data: {'mockResult': 0}, resolveToConfig: resolve_to_config,
         keepAlive: true});
    assert_true(validateSelectURLResult(select_url_result0, resolve_to_config));
    attachFencedFrame(select_url_result0, 'opaque-ads');
    const result0 = await nextValueFromServer(ancestor_key);
    assert_equals(result0, "frame0_loaded");

    let select_url_result1 = await sharedStorage.selectURL(
        "test-url-selection-operation", [{url: url0}, {url: url1}],
        {data: {'mockResult': 1}, resolveToConfig: resolve_to_config,
         keepAlive: true});
    assert_true(validateSelectURLResult(select_url_result1, resolve_to_config));
    attachFencedFrame(select_url_result1, 'opaque-ads');
    const result1 = await nextValueFromServer(ancestor_key);
    assert_equals(result1, "frame1_loaded");

    let select_url_result2 = await sharedStorage.selectURL(
        "test-url-selection-operation", [{url: url0}, {url: url1}],
        {data: {'mockResult': -1}, resolveToConfig: resolve_to_config,
         keepAlive: true});
    assert_true(validateSelectURLResult(select_url_result2, resolve_to_config));
    attachFencedFrame(select_url_result2, 'opaque-ads');
    const result2 = await nextValueFromServer(ancestor_key);
    assert_equals(result2, "frame0_loaded");
  }, 'selectURL() resolves to ' + (resolve_to_config ? 'config' : 'urn:uuid'));
}

</script>
</body>
